chore: refactor `CustomSoundsTable`
