+++ apps/meteor/client/views/admin/customSounds/CustomSoundsTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CustomSoundsTable';
